requirements.txt
flask
selenium
requests
assemblyai
google-generativeai
